Restructure the project
